requirements.txt
tatc
pydantic
mcp
pandas
geopandas
requests
python-dotenv
google-genai
google-generativeai